Moving storage connection tests over for JSONConnection.
- Had to add a mock class with API_URL_TEMPLATE, API_BASE_URL,
  API_VERSION set so we could test some methods.
- Removed unused imports in storage.connection module
- Rewrote some 'from httplib2 import Http' imports to avoid
  lint errors.
